parseScaleString is now a WIP
Currently: splits into an arraylist by-operator and then runs basic
evaluation to ensure a valid scale was given; throwing an invalid arg
exception if necessary. Also includes an overloaded findInArray function
which can search a specified range in an array for a character and the
index of that character is returned. (first find only which is why there
is a method to specify start index for easily finding a right bracket
for a given left etc etc)